Combine IFeatureActiveService & IFeatureDisableService to IFeatureManagementService
