+++ src/AElf.Kernel.FeatureManagement.Core/AElf.Kernel.FeatureManagement.Core.csproj
