+++ src/AElf.Kernel.FeatureManagement/AElf.Kernel.FeatureManagement.csproj
@@ -3,14 +3,14 @@
     <PropertyGroup>
         <TargetFramework>net8.0</TargetFramework>
         <LangVersion>latest</LangVersion>
-        <PackageId>AElf.Kernel.FeatureManager</PackageId>
+        <PackageId>AElf.Kernel.FeatureManagement</PackageId>
         <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
         <Description>Module that contains code to enable or disable new features of AElf.</Description>
     </PropertyGroup>
 
     <ItemGroup>
         <ProjectReference Include="..\AElf.Kernel.Configuration\AElf.Kernel.Configuration.csproj" />
-        <ProjectReference Include="..\AElf.Kernel.FeatureManager.Core\AElf.Kernel.FeatureManager.Core.csproj" />
+        <ProjectReference Include="..\AElf.Kernel.FeatureManagement.Core\AElf.Kernel.FeatureManagement.Core.csproj" />
     </ItemGroup>
 
 </Project>
