+++ src/AElf.Kernel.FeatureManagement/FeatureManagementOptions.cs
@@ -1,0 +1,8 @@
+using System.Collections.Generic;
+
+namespace AElf.Kernel.FeatureManagement;
+
+public class FeatureManagementOptions
+{
+    public List<string> DisableFeatureNameList { get; set; }
+}
