--- src/AElf.Kernel.FeatureManager.Core/IFeatureActiveService.cs
@@ -1,8 +1,0 @@
-using System.Threading.Tasks;
-
-namespace AElf.Kernel.FeatureManager;
-
-public interface IFeatureActiveService
-{
-    Task<bool> IsFeatureActive(string featureName);
-}
