--- src/AElf.Kernel.FeatureManager/DisableFeatureOptions.cs
@@ -1,8 +1,0 @@
-using System.Collections.Generic;
-
-namespace AElf.Kernel.FeatureManager;
-
-public class DisableFeatureOptions
-{
-    public List<string> FeatureNameList { get; set; }
-}
