+++ src/AElf.Kernel.SmartContract/SmartContractAElfModule.cs
@@ -1,4 +1,5 @@
-using AElf.Kernel.FeatureManager.Core;
+using AElf.Kernel.FeatureManagement;
+using AElf.Kernel.FeatureManagement.Core;
 using AElf.Kernel.SmartContract.Application;
 using AElf.Kernel.SmartContract.Infrastructure;
 using AElf.Modularity;
@@ -8,7 +9,7 @@ using Volo.Abp.Modularity;
 namespace AElf.Kernel.SmartContract;
 
 [DependsOn(typeof(CoreKernelAElfModule),
-    typeof(FeatureManagerCoreAElfModule))]
+    typeof(FeatureManagementCoreAElfModule))]
 public class SmartContractAElfModule : AElfModule
 {
     public override void ConfigureServices(ServiceConfigurationContext context)
