+++ test/AElf.Kernel.FeatureManagement.Tests/FeatureManagementTestModule.cs
@@ -5,12 +5,12 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Volo.Abp.Modularity;
 
-namespace AElf.Kernel.FeatureManager.Tests;
+namespace AElf.Kernel.FeatureManagement.Tests;
 
 [DependsOn(typeof(ContractTestModule),
-    typeof(FeatureManagerAElfModule),
+    typeof(FeatureManagementAElfModule),
     typeof(KernelAElfModule))]
-public class FeatureManagerTestModule : ContractTestModule
+public class FeatureManagementTestModule : ContractTestModule
 {
     public override void ConfigureServices(ServiceConfigurationContext context)
     {
