+++ test/AElf.Kernel.FeatureManagement.Tests/MinerList.cs
