+++ test/AElf.Kernel.SmartContract.ExecutionPluginForMethodFee.Tests/Service/PlainTransactionExecutingAsPluginService.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
-using AElf.Kernel.FeatureDisable.Core;
+using AElf.Kernel.FeatureManagement.Core;
 using AElf.Kernel.SmartContract.Application;
 using AElf.Kernel.SmartContract.Domain;
 using AElf.Kernel.SmartContract.Infrastructure;
@@ -21,8 +21,8 @@ public class PlainTransactionExecutingAsPluginService : PlainTransactionExecutin
     public PlainTransactionExecutingAsPluginService
     (ISmartContractExecutiveService smartContractExecutiveService,
         IEnumerable<IPostExecutionPlugin> postPlugins, IEnumerable<IPreExecutionPlugin> prePlugins,
-        ITransactionContextFactory transactionContextFactory, IFeatureDisableService featureDisableService) : base(
-        smartContractExecutiveService, postPlugins, prePlugins, transactionContextFactory, featureDisableService)
+        ITransactionContextFactory transactionContextFactory, IFeatureManagementService featureManagementService) : base(
+        smartContractExecutiveService, postPlugins, prePlugins, transactionContextFactory, featureManagementService)
     {
         _smartContractExecutiveService = smartContractExecutiveService;
     }
